# Avoid returning slice into buffer from Readfile