@shirou 2021-02-20
`process` package has a `BenchmarkProcessName`. So I did 

```
GODEBUG=gctrace=1  go test -bench BenchmarkProcessName  -benchmem
```

before

```
pkg: github.com/shirou/gopsutil/process
BenchmarkProcessName-2          gc 4 @1.765s 0%: 0.11+0.37+0.003 ms clock, 0.22+0/0.13/0.22+0.006 ms cpu, 0->0->0 MB, 4 MB goal, 2 P (forced)
gc 5 @1.767s 0%: 0.060+0.17+0.001 ms clock, 0.12+0/0.18/0.093+0.003 ms cpu, 0->0->0 MB, 4 MB goal, 2 P (forced)
gc 6 @1.767s 0%: 0.090+0.17+0.002 ms clock, 0.18+0/0.16/0.11+0.004 ms cpu, 0->0->0 MB, 4 MB goal, 2 P (forced)
gc 7 @1.777s 0%: 0.34+0.73+0.001 ms clock, 0.69+0/0.34/0.33+0.003 ms cpu, 0->0->0 MB, 4 MB goal, 2 P (forced)
gc 8 @2.357s 0%: 0.041+3.5+0.002 ms clock, 0.082+0/0.19/3.4+0.004 ms cpu, 0->0->0 MB, 4 MB goal, 2 P (forced)
207386386                5.76 ns/op            0 B/op          0 allocs/op
```

PR applied

```
pkg: github.com/shirou/gopsutil/process
BenchmarkProcessName-2          gc 4 @1.767s 0%: 0.23+0.24+0.002 ms clock, 0.47+0/0.29/0.078+0.004 ms cpu, 0->0->0 MB, 4 MB goal, 2 P (forced)
gc 5 @1.768s 0%: 0.050+0.16+0.001 ms clock, 0.10+0/0.15/0.11+0.003 ms cpu, 0->0->0 MB, 4 MB goal, 2 P (forced)
gc 6 @1.768s 0%: 0.13+0.34+0.001 ms clock, 0.26+0/0.082/0.19+0.003 ms cpu, 0->0->0 MB, 4 MB goal, 2 P (forced)
gc 7 @1.778s 0%: 0.12+0.27+0.002 ms clock, 0.25+0/0.16/0.25+0.005 ms cpu, 0->0->0 MB, 4 MB goal, 2 P (forced)
gc 8 @2.356s 0%: 0.23+0.56+0.005 ms clock, 0.46+0/0.56/0.12+0.011 ms cpu, 0->0->0 MB, 4 MB goal, 2 P (forced)
207829984                6.10 ns/op            0 B/op          0 allocs/op
```

Seems nothing changed. And I try to change get new Process() every time, also nothing changed.

```
 func BenchmarkProcessName(b *testing.B) {
-       p := testGetProcess()
        for i := 0; i < b.N; i++ {
+               p := testGetProcess()
                p.Name()
        }
 }
```

Am I missed something?