@shirou 2021-02-26
Understand. This is realloc issue, so heap alloc keep low but GC is much more. Keep heap lower is better, so I merge this PR. Thank you for sharing your cool knowledge about memory!